% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AndrewsWeightFunction.R
\name{AndrewsWeightFunction}
\alias{AndrewsWeightFunction}
\title{Andrews-sine weight function}
\usage{
AndrewsWeightFunction(x, q)
}
\description{
Applies weights to the observations based on the Andrews-sine weight function
}
