% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussWeightFunction.R
\name{GaussWeightFunction}
\alias{GaussWeightFunction}
\title{Generalized-Gauss weight function}
\usage{
GaussWeightFunction(x, q, a, b)
}
\description{
Applies weights to the observations based on the Generalized-Gauss weight function
}
