% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HampelWeightFunction.R
\name{HampelWeightFunction}
\alias{HampelWeightFunction}
\title{Hampel weight function}
\usage{
HampelWeightFunction(x, q1, q2, q3)
}
\description{
Applies weights to the observations based on the Hampel weight function
}
