% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HuberWeightFunction.R
\name{HuberWeightFunction}
\alias{HuberWeightFunction}
\title{Huber weight function}
\usage{
HuberWeightFunction(x, q)
}
\description{
Applies weights to the observations based on the Huber weight function
}
