% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuadraticWeightFunction.R
\name{QuadraticWeightFunction}
\alias{QuadraticWeightFunction}
\title{Linear quadratic quadratic weight function}
\usage{
QuadraticWeightFunction(x, q1, q2, s)
}
\description{
Applies weights to the observations based on the linear quadratic quadratic weight function
}
