% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleResidualsByMAD.R
\name{scaleResidualsByMAD}
\alias{scaleResidualsByMAD}
\title{Residual scaling}
\usage{
scaleResidualsByMAD(residuals)
}
\arguments{
\item{residuals}{the residuals.}
}
\value{
the scaled residuals.
}
\description{
Residual scaling
}
