% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TukeyWeightFunction.R
\name{TukeyWeightFunction}
\alias{TukeyWeightFunction}
\title{Tukey's bisquare weight function}
\usage{
TukeyWeightFunction(x, q)
}
\description{
Applies weights to the observations based on the Tukey's bisquare weight function
}
