% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_functional.R
\name{crm_functional}
\alias{crm_functional}
\title{Cellwise Robust M regression with additional input arguments}
\usage{
crm_functional(
  formula,
  data,
  maxiter = 100,
  tolerance = 0.01,
  outlyingness.factor = 1,
  spadieta = seq(0.9, 0.1, -0.1),
  center = "median",
  scale = "qn",
  regtype = "MM",
  alphaLTS = NULL,
  seed = NULL,
  verbose = TRUE,
  weightfunction = "hampel",
  weightthreshold = 1,
  parameters
)
}
\description{
Estimates coefficients by applying Cellwise Robust M regression
}
