% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_outlying_cells.R
\name{impute_outlying_cells}
\alias{impute_outlying_cells}
\title{Imputation of outlying cells}
\usage{
impute_outlying_cells(
  outlier,
  data,
  outlvarz,
  az,
  wr,
  increment = 0,
  threshold
)
}
\description{
Imputation of outlying cells
}
